add shell model
