+++ types/enum/element.ts
@@ -1,3 +1,21 @@
+export enum ElementSymbol {
+  H = 'h', He = 'he', Li = 'li', Be = 'be', B = 'b', C = 'c', N = 'n', O = 'o', F = 'f',
+  Ne = 'ne', Na = 'na', Mg = 'mg', Al = 'al', Si = 'si', P = 'p', S = 's', Cl = 'cl',
+  Ar = 'ar', K = 'k', Ca = 'ca', Sc = 'sc', Ti = 'ti', V = 'v', Cr = 'cr', Mn = 'mn',
+  Fe = 'fe', Co = 'co', Ni = 'ni', Cu = 'cu', Zn = 'zn', Ga = 'ga', Ge = 'ge', As = 'as',
+  Se = 'se', Br = 'br', Kr = 'kr', Rb = 'rb', Sr = 'sr', Y = 'y', Zr = 'zr', Nb = 'nb',
+  Mo = 'mo', Tc = 'tc', Ru = 'ru', Rh = 'rh', Pd = 'pd', Ag = 'ag', Cd = 'cd', In = 'in',
+  Sn = 'sn', Sb = 'sb', Te = 'te', I = 'i', Xe = 'xe', Cs = 'cs', Ba = 'ba', La = 'la',
+  Ce = 'ce', Pr = 'pr', Nd = 'nd', Pm = 'pm', Sm = 'sm', Eu = 'eu', Gd = 'gd', Tb = 'tb',
+  Dy = 'dy', Ho = 'ho', Er = 'er', Tm = 'tm', Yb = 'yb', Lu = 'lu', Hf = 'hf', Ta = 'ta',
+  W = 'w', Re = 're', Os = 'os', Ir = 'ir', Pt = 'pt', Au = 'au', Hg = 'hg', Tl = 'tl',
+  Pb = 'pb', Bi = 'bi', Po = 'po', At = 'at', Rn = 'rn', Fr = 'fr', Ra = 'ra', Ac = 'ac',
+  Th = 'th', U = 'u', Np = 'np', Pu = 'pu', Am = 'am', Cm = 'cm', Bk = 'bk', Cf = 'cf',
+  Es = 'es', Fm = 'fm', Md = 'md', No = 'no', Lr = 'lr', Rf = 'rf', Db = 'db', Sg = 'sg',
+  Bh = 'bh', Hs = 'hs', Mt = 'mt', Ds = 'ds', Rg = 'rg', Cn = 'cn', Nh = 'nh', Fl = 'fl',
+  Mc = 'mc', Lv = 'lv', Ts = 'ts', Og = 'og'
+};
+
 export enum ElementBlock { S = 's', P = 'p', D = 'd', F = 'f' };
 
 export enum ElementSet {
