+++ types/enum/generic.ts
@@ -9,24 +9,6 @@ export enum LangCode {
   THAI = 'th', TURKISH = 'tr', UKRAINIAN = 'uk', VIETNAMESE = 'vi'
 };
 
-export enum ElementSymbol {
-  H = 'h', He = 'he', Li = 'li', Be = 'be', B = 'b', C = 'c', N = 'n', O = 'o', F = 'f',
-  Ne = 'ne', Na = 'na', Mg = 'mg', Al = 'al', Si = 'si', P = 'p', S = 's', Cl = 'cl',
-  Ar = 'ar', K = 'k', Ca = 'ca', Sc = 'sc', Ti = 'ti', V = 'v', Cr = 'cr', Mn = 'mn',
-  Fe = 'fe', Co = 'co', Ni = 'ni', Cu = 'cu', Zn = 'zn', Ga = 'ga', Ge = 'ge', As = 'as',
-  Se = 'se', Br = 'br', Kr = 'kr', Rb = 'rb', Sr = 'sr', Y = 'y', Zr = 'zr', Nb = 'nb',
-  Mo = 'mo', Tc = 'tc', Ru = 'ru', Rh = 'rh', Pd = 'pd', Ag = 'ag', Cd = 'cd', In = 'in',
-  Sn = 'sn', Sb = 'sb', Te = 'te', I = 'i', Xe = 'xe', Cs = 'cs', Ba = 'ba', La = 'la',
-  Ce = 'ce', Pr = 'pr', Nd = 'nd', Pm = 'pm', Sm = 'sm', Eu = 'eu', Gd = 'gd', Tb = 'tb',
-  Dy = 'dy', Ho = 'ho', Er = 'er', Tm = 'tm', Yb = 'yb', Lu = 'lu', Hf = 'hf', Ta = 'ta',
-  W = 'w', Re = 're', Os = 'os', Ir = 'ir', Pt = 'pt', Au = 'au', Hg = 'hg', Tl = 'tl',
-  Pb = 'pb', Bi = 'bi', Po = 'po', At = 'at', Rn = 'rn', Fr = 'fr', Ra = 'ra', Ac = 'ac',
-  Th = 'th', U = 'u', Np = 'np', Pu = 'pu', Am = 'am', Cm = 'cm', Bk = 'bk', Cf = 'cf',
-  Es = 'es', Fm = 'fm', Md = 'md', No = 'no', Lr = 'lr', Rf = 'rf', Db = 'db', Sg = 'sg',
-  Bh = 'bh', Hs = 'hs', Mt = 'mt', Ds = 'ds', Rg = 'rg', Cn = 'cn', Nh = 'nh', Fl = 'fl',
-  Mc = 'mc', Lv = 'lv', Ts = 'ts', Og = 'og'
-};
-
 export enum PTColumn {
   COLUMN_1 = 1, COLUMN_2 = 2, COLUMN_3 = 3, COLUMN_4 = 4, COLUMN_5 = 5, COLUMN_6 = 6,
   COLUMN_7 = 7, COLUMN_8 = 8, COLUMN_9 = 9, COLUMN_10 = 10, COLUMN_11 = 11, COLUMN_12 = 12,
@@ -38,6 +20,8 @@ export enum PTPeriod {
   PERIOD_1 = 1, PERIOD_2 = 2, PERIOD_3 = 3, PERIOD_4 = 4, PERIOD_5 = 5, PERIOD_6 = 6,
   PERIOD_7 = 7
 };
+
+export enum ShellModel { K = 'k', L = 'l', M = 'm', N = 'n', O = 'o', P = 'p', Q = 'q' };
 
 export enum Phase {
   SOLID = 'solid',
